requirements/prod.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/prod.txt requirements/prod.in
#
alembic==1.13.1
    # via -r requirements/prod.in
annotated-types==0.6.0
    # via pydantic
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via -r requirements/prod.in
greenlet==3.0.3
    # via sqlalchemy
idna==3.7
    # via email-validator
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
psycopg2==2.9.9
    # via -r requirements/prod.in
pydantic==2.7.1
    # via -r requirements/prod.in
pydantic-core==2.18.2
    # via pydantic
python-ulid==2.6.0
    # via -r requirements/prod.in
sqlalchemy==2.0.30
    # via
    #   -r requirements/prod.in
    #   alembic
typing-extensions==4.11.0
    # via
    #   alembic
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
